Add program to send an email whenever anyone logs in
The program is designed to send an email to a cell-phone whenever
any user logs into any of my machines.  If the user is allowed to
log in, they will know to activate the counter-program that stops
the email from going out.

Tentative ideas for integration involve putting a call to this
script in the global .bashrc or global .login file.

NOT FINISHED