Adds mail archiving script
This script is being developed on SDF.  It gets copied over to this
repo b/c there's no git support (that i have access to) on SDF.

This utility will go through my messages folder and archive unsorted
messages that are older that a certain date into "archive" folders.

I suppose this could be made to work inside the sorted messages'
folders as well.