Create tests_passed folder. Finish offlimit program.
Create a folder for successful experiments as a way to get
non-essential scripts out of the way.

Also, finished the prgm_offlimits.pl script.  Now I can make an
alias for "forbidden" programs to include this script before
their normal execution.  I may modify a way to configure the
hours.  (Not sure how this will work with things like Ubuntu's
shortcuts on the sidebars)