Created BSD folder. Made simple auto-email script.
Created the BSD folder to contain a bunch of command and response
scripts that will be used for interacting with myself via phone
from across the internet.

The first script was a simple email that sends the results of the
"uptime" command to the email address listed in the config file.