Add watchcat.pl
Began program watchcat.  When run with no arguments, it sleeps for a short
period of time and then kills the process of the host tty.  The user needs
to run the same program with the "feed" argument in a short period of time
in order to prevent that from happening.

This is a security program.

As of right now, the functionality to kill to the process does not exist.
Currently only displays messages and makes a log