Added rsync_test.pl. Runs batches of rsync tests.
Created a file that will run a bunch of batched rsync tests. It will
create initial files, run rsync with different options, and let you
examine the results.  This is part of a larger effort to try to
interpret rsync errors.

The first test is completed. It creates folders and files.  It runs a
common set of options for rsync (-av).  There is no logging or error
message checking.  There is no run-time error trapping. Nothing is
deleted after the test is run.