Fixed a bug with rsync test 01 and successfully ran
Fixed a bug where we were trying to explicitly change directories
to the parent directory, but the string stored in memory did not
include the full path.

Successfully ran the test. The perl test-driver program doesn't output
anything, but rsync itself does.  Do we want to consider sending that output
somewhere else? Or maybe turn off the -v flag in rsync?