Copy dirs with slash. Tested Extra Files
Original test specified directories without the
trailing slash and it copied the folder in
addition to the files.

Added test that included the trailing slash, and
only the files inside were copied.

Added tests where there was an extra file in the
destination folder. It was not removed during the
rsync call.