Add script to create multiple test webpages
Created a script that will create plain, mostly-empty, test
web pages based on a list of names.

Also modified the script that makes index.html pages to include
directories and to attempt to exclude certain links. Was partially
successful.  It would exclude it's own name, but it wouldn't exclude
the entries "." and "..".