Exif-tag script now produces single results file
Modified gather_media_info.pl so that it returns some data objects
(list of lists).  This way, all the information can be compiled into
one master file for the root directory.

Also modified the program to put in placeholder text when a file
doesn't have one of the desired tags.