Begin basic Tk image viewer
In preparation for doing more image manipulation in Perl, I'm trying to
set up a good foundation for easily running experiments and tests.  I'm
currently able to display the image, but i'm trying to figure out how to
shrink it down proportionally.  The example that I chose is a bit big.

Script is not complete.